# examples/transmitter/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Stream Management API for OpenID Shared Security Events
  description: "[OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)\n\
    \nHTTP API to be implemented by Event Transmitters. This API can be used by\n\
    Event Receivers to query and update the Event Stream configuration and\nstatus,\
    \ to add and remove subjects, and to trigger verification.\n"
  license:
    name: MIT
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /.well-known/sse-configuration/{issuer}:
    get:
      tags:
      - StreamManagement
      summary: Transmitter Configuration Request (with path)
      description: Return Transmitter Configuration information (with support for
        specifying an issuer).
      operationId: _well_known_sse_configuration_issuer_get
      parameters:
      - name: issuer
        in: path
        description: "Using path components enables supporting multiple issuers per\
          \ host.\nThis is required in some multi-tenant hosting configurations.\n\
          This use of .well-known is for supporting multiple issuers per host;\nunlike\
          \ its use in [RFC5785](https://openid.net/specs/openid-sse-framework-1_0.html#RFC5785),\n\
          it does not provide general information about the host.\n"
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: tenant-a
      responses:
        "200":
          description: "A set of Claims about the Transmitter's configuration,\nincluding\
            \ all necessary endpoints and public key location information.\n"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransmitterConfiguration'
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
  /stream:
    get:
      tags:
      - StreamManagement
      summary: Request to retrieve the configuration of an event stream
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-configuration)\n\
        \nAn Event Receiver gets the current configuration of a stream by making an\
        \ HTTP GET request to the Configuration\nEndpoint. On receiving a valid request\
        \ the Event Transmitter responds with a 200 OK response containing a JSON\n\
        representation of the stream’s configuration in the body."
      operationId: stream_get
      responses:
        "200":
          description: The Event Stream's configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
        "401":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to read the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: There is no Event Stream configured for this Event Receiver.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
    post:
      tags:
      - StreamManagement
      summary: Request to update the configuration of an event stream
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#updating-a-streams-configuration)\n\
        \nAn Event Receiver updates the current configuration of a stream by making\
        \ an HTTP POST request to the Configuration\nEndpoint. The POST body contains\
        \ a JSON representation of the updated configuration. On receiving a valid\
        \ request\nthe Event Transmitter responds with a 200 OK response containing\
        \ a JSON representation of the updated stream\nconfiguration in the body.\n\
        \nThe full set of editable properties must be present in the POST body, not\
        \ only the ones that are specifically\nintended to be changed. Missing properties\
        \ SHOULD be interpreted as requested to be deleted. Event Receivers should\n\
        read the configuration first, modify the JSON representation, then make an\
        \ update request.\n\nProperties that cannot be updated MAY be present, but\
        \ they MUST match the expected value."
      operationId: stream_post
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamConfiguration'
        required: true
      responses:
        "200":
          description: The Event Stream's updated configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
        "202":
          description: "The update request has been accepted, but not processed.\n\
            Receiver MAY try the same request later in order to get processing result.\n\
            \nIf a Receiver makes a request to update the stream to enable it for\
            \ a specific Subject, and the Transmitter is\nunable to decide whether\
            \ or not to complete the request, then the Transmitter MUST respond with\
            \ a 202 status\ncode."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: There is no Event Stream configured for this Event Receiver.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
    delete:
      tags:
      - StreamManagement
      summary: Request to remove the configuration of an event stream
      description: |-
        An Event Receiver removes the configuration of a stream by making an HTTP DELETE request to the Configuration
        Endpoint. On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
        successfully removed.
      operationId: stream_delete
      responses:
        "200":
          description: |-
            On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
            successfully removed.
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
  /status:
    get:
      tags:
      - StreamManagement
      summary: Request to get the status of an Event Stream
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-status)\n\
        \nAn Event Receiver checks the current status of an event stream by making\
        \ an\nHTTP GET request to the stream’s Status Endpoint."
      operationId: get_status
      parameters:
      - name: subject
        in: query
        description: |-
          OPTIONAL. The subject for which the stream status is requested.
          Content is a string-encoded json representation of a Subject
          (see Subject schema)
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "{\"user\":{\"format\":\"email\",\"email\":\"reginold@popular-app.com\"\
            }}"
      responses:
        "200":
          description: |-
            On receiving a valid request the Event Transmitter responds with a 200
            OK response containing a JSON object with an attribute status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamStatus'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to read this stream's status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: "There is no Event Stream configured for this Event Receiver,\n\
            or the Subject specified is invalid or\nthe Receiver is not authorized\
            \ to get status for the specified Subject."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
    post:
      tags:
      - StreamManagement
      summary: Request to update an Event Stream's status
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0-ID1.html#updating-a-streams-status)\n\
        \nAn Event Receiver updates the current status of a stream by making an HTTP\n\
        POST request to the Status Endpoint."
      operationId: update_status
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStreamStatus'
        required: true
      responses:
        "200":
          description: Event Stream status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStreamStatus'
        "202":
          description: "The update request has been accepted, but not processed.\n\
            Receiver MAY try the same request later in order to get processing result."
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: The Event Receiver is not allowed to update the stream status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: "There is no Event Stream configured for this Event Receiver,\n\
            or an invalid Subject is specified."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
  /add-subject:
    post:
      tags:
      - StreamManagement
      summary: Request to add a subject to an Event Stream
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#adding-a-subject-to-a-stream)\n\
        \nEvent Receivers can send requests to an Event Transmitter's Add Subject\
        \ endpoint to add a subject to an Event Stream."
      operationId: add_subject
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSubjectParameters'
        required: true
      responses:
        "200":
          description: "On a successful response, the Event Transmitter responds with\
            \ an empty 200 OK response.\nThe Event Transmitter MAY choose to silently\
            \ ignore the request, for example if the subject has previously\nindicated\
            \ to the transmitter that they do not want events to be transmitted to\
            \ the Event Receiver.\nIn this case, the transmitter MAY return an empty\
            \ 200 OK response or an appropriate error code.\nSee [Security Considerations](https://openid.net/specs/openid-sse-framework-1_0.html#management-sec)."
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: "The Event Receiver is not allowed to add this particular subject,\
            \ or not allowed to add in general."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: "The subject is not recognized by the Event Transmitter,\n\
            the Event Transmitter may chose to stay silent in this case and respond\
            \ with 200."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
  /remove-subject:
    post:
      tags:
      - StreamManagement
      summary: Request to add a subject to an Event Stream
      description: "[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#removing-a-subject)\n\
        \nEvent Receivers can send requests to an Event Transmitter's Remove Subject\
        \ endpoint to remove a subject from an Event Stream."
      operationId: remove_subject
      requestBody:
        description: Request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveSubjectParameters'
        required: true
      responses:
        "204":
          description: "On a successful response, the Event Transmitter responds with\
            \ a 204 No Content response."
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: "The Event Receiver is not allowed to remove this particular\
            \ subject, or not allowed to remove in general."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: "The subject is not recognized by the Event Transmitter,\n\
            the Event Transmitter may chose to stay silent in this case and respond\
            \ with 204."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
  /verification:
    post:
      tags:
      - StreamManagement
      summary: Request that a verification event be sent over an Event Stream
      operationId: verification_request
      requestBody:
        description: Optional request parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationParameters'
        required: false
      responses:
        "204":
          description: "Request for verification event was successfully received.\
            \ A\nsuccessful response from a POST to the Verification Endpoint does\n\
            not indicate that the verification event was transmitted\nsuccessfully,\
            \ only that the Event Transmitter has transmitted the\nevent or will do\
            \ so at some point in the future.\n"
        "400":
          description: Request body cannot be parsed or the request is otherwise invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authorization failed or is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Event Receiver is sending too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.stream_management_controller
...
components:
  schemas:
...
    TransmitterConfiguration:
      title: Transmitter Configuration Response
      required:
      - issuer
      - jwks_uri
      type: object
      properties:
        issuer:
          type: string
          description: |
            URL using the https scheme with no query or fragment component that the Transmitter asserts as its
            Issuer Identifier.
            This MUST be identical to the iss claim value in Security Event Tokens issued from this Transmitter.
          format: uri
          example: https://most-secure.com
        jwks_uri:
          type: string
          description: "URL of the Transmitter's [JSON Web Key Set](https://openid.net/specs/openid-sse-framework-1_0.html#RFC7517)\n\
            document. This contains the signing key(s) the Receiver uses to validate\
            \ signatures from the Transmitter.\n"
          format: uri
          example: https://transmitter.most-secure.com/jwks.json
        delivery_methods_supported:
          type: array
          description: List of supported delivery method URIs. Recommended.
          example:
          - https://schemas.openid.net/secevent/risc/delivery-method/push
          - https://schemas.openid.net/secevent/risc/delivery-method/poll
          items:
            type: string
            format: uri
        configuration_endpoint:
          type: string
          description: The URL of the Configuration Endpoint.
          format: uri
          example: https://transmitter.most-secure.com/stream
        status_endpoint:
          type: string
          description: The URL of the Status Endpoint.
          format: uri
          example: https://transmitter.most-secure.com/status
        add_subject_endpoint:
          type: string
          description: The URL of the Add Subject Endpoint.
          format: uri
          example: https://transmitter.most-secure.com/add-subject
        remove_subject_endpoint:
          type: string
          description: The URL of the Remove Subject Endpoint.
          format: uri
          example: https://transmitter.most-secure.com/remove-subject
        verification_endpoint:
          type: string
          description: The URL of the Verification Endpoint.
          format: uri
          example: https://transmitter.most-secure.com/verification
        critical_subject_members:
          type: array
          description: "List of member names in a Complex Subject which, if present\
            \ in a Subject Member in an event,\nMUST be interpreted by a Receiver.\n"
          example:
          - tenant
          - user
          items:
            type: string
      description: "Transmitters have metadata describing their configuration.\n[OpenID\
        \ Spec](https://openid.net/specs/openid-sse-framework-1_0.html#discovery-meta)\n"
      example:
        delivery_methods_supported:
        - https://schemas.openid.net/secevent/risc/delivery-method/push
        - https://schemas.openid.net/secevent/risc/delivery-method/poll
        configuration_endpoint: https://transmitter.most-secure.com/stream
        critical_subject_members:
        - tenant
        - user
        add_subject_endpoint: https://transmitter.most-secure.com/add-subject
        remove_subject_endpoint: https://transmitter.most-secure.com/remove-subject
        status_endpoint: https://transmitter.most-secure.com/status
        jwks_uri: https://transmitter.most-secure.com/jwks.json
        issuer: https://most-secure.com
        verification_endpoint: https://transmitter.most-secure.com/verification
    StreamConfiguration:
      title: Event Stream Configuration
      required:
      - delivery
      - events_requested
      type: object
      properties:
        iss:
          type: string
          description: |-
            Read-Only.
            A URL using the https scheme with no query or fragment component that the Transmitter asserts as its Issuer
            Identifier. This MUST be identical to the iss Claim value in Security Event Tokens issued from this Transmitter.
          format: uri
          example: https://most-secure.com
        aud:
          description: "Read-Only.\nA string or an array of strings containing an\
            \ audience claim as defined in\n[JSON Web Token (JWT)](https://openid.net/specs/openid-sse-framework-1_0.html#RFC7519)\
            \ that identifies\nthe Event Receiver(s) for the Event Stream. This property\
            \ cannot be updated. If multiple Receivers are specified\nthen the Transmitter\
            \ SHOULD know that these Receivers are the same entity."
          example: https://popular-app.com
          oneOf:
          - type: string
            format: uri
          - type: array
            items:
              type: string
              format: uri
        events_supported:
          type: array
          description: "Read-Only.\nAn array of URIs identifying the set of events\
            \ supported by the Transmitter for this Receiver.\nIf omitted, Event Transmitters\
            \ SHOULD make this set available to the Event Receiver via some other\
            \ means\n(e.g. publishing it in online documentation)."
          example:
          - https://schemas.openid.net/secevent/caep/event-type/session-revoked
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        events_requested:
          type: array
          description: |-
            Read-Write.
            An array of URIs identifying the set of events that the Receiver requested.
            A Receiver SHOULD request only the events that it understands and it can act on.
            This is configurable by the Receiver.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        events_delivered:
          type: array
          description: |-
            Read-Only.
            An array of URIs which is the intersection of events_supported and events_requested.
            These events MAY be delivered over the Event Stream.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            type: string
            format: uri
        delivery:
          description: "Read-Write.\nA JSON object containing a set of name/value\
            \ pairs specifying configuration parameters for the SET delivery\nmethod.\
            \ The actual delivery method is identified by the special key method with\
            \ the value being a URI as defined\nin [Section 11.2.1](https://openid.net/specs/openid-sse-framework-1_0.html#delivery-meta)."
          example:
            method: https://schemas.openid.net/secevent/risc/delivery-method/poll
            endpoint_url: null
          discriminator:
            propertyName: method
          oneOf:
          - $ref: '#/components/schemas/PushDeliveryMethod'
          - $ref: '#/components/schemas/PollDeliveryMethod'
        min_verification_interval:
          type: integer
          description: "Read-Only.\nAn integer indicating the minimum amount of time\
            \ in seconds that must pass in between verification requests.\nIf an Event\
            \ Receiver submits verification requests more frequently than this, the\
            \ Event Transmitter MAY respond\nwith a 429 status code. An Event Transmitter\
            \ SHOULD NOT respond with a 429 status code if an Event Receiver is not\n\
            exceeding this frequency."
        format:
          type: string
          description: |-
            Read-Write.
            The Subject Identifier Format that the Receiver wants for the events.
            If not set then the Transmitter might decide to use a type that discloses more information than necessary.
      description: "JSON Object describing and Event Stream's configuration\n[Spec](https://openid.net/specs/openid-sse-framework-1_0.html#stream-config)\"\
        \n"
      example:
        aud: https://popular-app.com
        delivery:
          method: https://schemas.openid.net/secevent/risc/delivery-method/poll
          endpoint_url: null
        events_supported:
        - https://schemas.openid.net/secevent/caep/event-type/session-revoked
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        iss: https://most-secure.com
        format: format
        events_delivered:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        events_requested:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        min_verification_interval: 0
    UpdateStreamStatus:
      type: object
      properties:
        reason:
          type: string
          description: OPTIONAL. A short text description that explains the reason
            for the change.
          example: Disabled by administrator action.
      example:
        reason: Disabled by administrator action.
      allOf:
      - $ref: '#/components/schemas/StreamStatus'
...
    AddSubjectParameters:
      required:
      - subject
      type: object
      properties:
        subject:
          allOf:
          - $ref: '#/components/schemas/Subject'
          - description: REQUIRED. A Subject claim identifying the subject to be added.
        verified:
          type: boolean
          description: "OPTIONAL. A boolean value; when true, it indicates that the\
            \ Event Receiver has verified the Subject claim.\nWhen false, it indicates\
            \ that the Event Receiver has not verified the Subject claim.\nIf omitted,\
            \ Event Transmitters SHOULD assume that the subject has been verified."
      example:
        subject:
          format: email
          email: reginold@popular-app.com
        verified: true
...
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth